backend/requirements.txt
fastapi
uvicorn[standard]
supabase
openai
python-dotenv
pydantic
psycopg2-binary
qrcode[pil]

